Rename design experiment document
This document is really the description of the infrastructure. The authoritative source of experimental design is currently either in the original google doc or the paper currently being written. This document tries to be the infrastructure document and the experiment design at the same time which it doesn't succeed at and is confusing.